Sécurités empêchant accès fonctions admin

+++ src/controller/BackController.php
@@ -8,121 +8,164 @@ use Mich\Blog\config\Parameter;
 
 class BackController extends Controller
 {
+    // Vérifie si utilisateur est connecté
+    private function checkLoggedIn()
+    {
+        if (!$this->session->get("nickname")) {
+            $this->session->set("need_login", "Vous devez vous connecter pour accéder à cette page");
+            header("Location: ../public/index.php?route=login");
+        } else {
+            return true;
+        }
+    }
+
+    // Vérifie si utilisateur connecté est admin
+    private function checkAdmin()
+    {
+        $this->checkLoggedIn();
+        if (!($this->session->get("role") === "admin")) {
+            $this->session->set("not_admin", "Vous n'avez pas le droit d'accéder à cette page");
+            header("Location: ../public/index.php?route=profile");
+        } else {
+            return true;
+        }
+    }
+
     // Accès espace admin et y apparaissent tous les articles, commentaires, utilisateurs
     public function administration()
     {
-        $articles = $this->articleDAO->getArticles();
-        $comments = $this->commentDAO->getAllComments();
-        $users = $this->userDAO->getUsers();
+        if ($this->checkAdmin()) {
+            $articles = $this->articleDAO->getArticles();
+            $comments = $this->commentDAO->getAllComments();
+            $users = $this->userDAO->getUsers();
 
-        return $this->view->render("administration", [
-            "articles" => $articles,
-            "comments" => $comments,
-            "users" => $users
-        ]);
+            return $this->view->render("administration", [
+                "articles" => $articles,
+                "comments" => $comments,
+                "users" => $users
+            ]);
+        }
     }
 
     // Ajout d'un article
     public function addArticle(Parameter $post)
     {
-        if ($post->get("submit")) {
-            $errors = $this->validation->validate($post, "Article"); // Processus de validation des données
-            if (!$errors) {
-                $this->articleDAO->addArticle($post, $this->session->get("id")); // 2ème param récup id de l'user connecté (admin)
-                $this->session->set("add_article", "Le nouvel article a bien été ajouté");
-                header("Location: ../public/index.php?route=administration");
+        if ($this->checkAdmin()) {
+            if ($post->get("submit")) {
+                $errors = $this->validation->validate($post, "Article"); // Processus de validation des données
+                if (!$errors) {
+                    $this->articleDAO->addArticle($post, $this->session->get("id")); // 2ème param récup id de l'user connecté (admin)
+                    $this->session->set("add_article", "Le nouvel article a bien été ajouté");
+                    header("Location: ../public/index.php?route=administration");
+                }
+                return $this->view->render("add_article", [
+                    "post" => $post,
+                    "errors" => $errors
+                ]);
             }
-            return $this->view->render("add_article", [
-                "post" => $post,
-                "errors" => $errors
-            ]);
+            return $this->view->render("add_article");
         }
-        return $this->view->render("add_article");
     }
 
     // Modification d'un article
     public function editArticle(Parameter $post, $articleId)
     {
-        $article = $this->articleDAO->getArticle($articleId); // Récupère d'abord le contenu de l'article
+        if ($this->checkAdmin()) {
+            $article = $this->articleDAO->getArticle($articleId); // Récupère d'abord le contenu de l'article
 
-        if ($post->get("submit")) {
-            $errors = $this->validation->validate($post, "Article");
-            if (!$errors) {
-                $this->articleDAO->editArticle($post, $articleId, $this->session->get("id")); // 3ème param récup id user connecté (admin)
-                $this->session->set("edit_article", "L'article a bien été modifié");
-                header("Location: ../public/index.php?route=administration");
+            if ($post->get("submit")) {
+                $errors = $this->validation->validate($post, "Article");
+                if (!$errors) {
+                    $this->articleDAO->editArticle($post, $articleId, $this->session->get("id")); // 3ème param récup id user connecté (admin)
+                    $this->session->set("edit_article", "L'article a bien été modifié");
+                    header("Location: ../public/index.php?route=administration");
+                }
+                return $this->view->render("edit_article", [
+                    "post" => $post,
+                    "errors" => $errors
+                ]);
             }
+            $post->set("id", $article->getId());
+            $post->set("title", $article->getTitle());
+            $post->set("content", $article->getContent());
+            $post->set("author", $article->getAuthor());
+
             return $this->view->render("edit_article", [
-                "post" => $post,
-                "errors" => $errors
+                "post" => $post
             ]);
         }
-        $post->set("id", $article->getId());
-        $post->set("title", $article->getTitle());
-        $post->set("content", $article->getContent());
-        $post->set("author", $article->getAuthor());
-
-        return $this->view->render("edit_article", [
-            "post" => $post
-        ]);
     }
-    
+
     // Suppression d'un article
     public function deleteArticle($articleId)
     {
-        $this->articleDAO->deleteArticle($articleId);
-        $this->session->set("delete_article", "L'article a bien été supprimé");
-        header("Location: ../public/index.php?route=administration");
+        if ($this->checkAdmin()) {
+            $this->articleDAO->deleteArticle($articleId);
+            $this->session->set("delete_article", "L'article a bien été supprimé");
+            header("Location: ../public/index.php?route=administration");
+        }
     }
 
     // Suppression d'un commentaire
     public function deleteComment($commentId)
     {
-        $this->commentDAO->deleteComment($commentId);
-        $this->session->set("delete_comment", "Le commentaire a bien été supprimé");
-        header("Location: ../public/index.php?route=administration");
+        if ($this->checkAdmin()) {
+            $this->commentDAO->deleteComment($commentId);
+            $this->session->set("delete_comment", "Le commentaire a bien été supprimé");
+            header("Location: ../public/index.php?route=administration");
+        }
     }
 
     // Enlève le signalement commentaire via l'espace admin
     public function unflagComment($commentId)
     {
-        $this->commentDAO->unflagComment($commentId);
-        $this->session->set("unflag_comment", "Le commentaire n'est plus signalé");
-        header("Location: ../public/index.php?route=administration");
+        if ($this->checkAdmin()) {
+            $this->commentDAO->unflagComment($commentId);
+            $this->session->set("unflag_comment", "Le commentaire n'est plus signalé");
+            header("Location: ../public/index.php?route=administration");
+        }
     }
 
     // Accès au profil utilisateur
     public function profile()
     {
-        return $this->view->render("profile");
+        if ($this->checkLoggedIn()) {
+            return $this->view->render("profile");
+        }
     }
 
     // Changement du mot de passe utilisateur
     public function updatePassword(Parameter $post)
     {
-        if ($post->get("submit")) {
-            $this->userDAO->updatePassword($post, $this->session->get("nickname"));
-            $this->session->set("update_password", "Le mot de passe a été mis à jour");
-            header("Location: ../public/index.php?route=profile");
+        if ($this->checkLoggedIn()) {
+            if ($post->get("submit")) {
+                $this->userDAO->updatePassword($post, $this->session->get("nickname"));
+                $this->session->set("update_password", "Le mot de passe a été mis à jour");
+                header("Location: ../public/index.php?route=profile");
+            }
+            return $this->view->render("update_password");
         }
-        return $this->view->render("update_password");
     }
 
     // Déconnexion utilisateur
     public function logout()
     {
-        $this->logoutOrDelete("logout");
+        if ($this->checkLoggedIn()) {
+            $this->logoutOrDelete("logout");
+        }
     }
 
     // Suppression compte utilisateur
     public function deleteAccount()
     {
-        $this->userDAO->deleteAccount($this->session->get("nickname"));
-        $this->logoutOrDelete("delete_account");
+        if ($this->checkLoggedIn()) {
+            $this->userDAO->deleteAccount($this->session->get("nickname"));
+            $this->logoutOrDelete("delete_account");
+        }
     }
 
     // Méthodes logout() et deleteAccount() similaires, évite la répétition de code dans ces deux méthodes
-    public function logoutOrDelete($param)
+    private function logoutOrDelete($param)
     {
         $this->session->stop();
         $this->session->start();
@@ -137,8 +180,10 @@ class BackController extends Controller
     // Suppression utilisateur depuis espace admin
     public function deleteUser($userId)
     {
-        $this->userDAO->deleteUser($userId);
-        $this->session->set("delete_user", "L'utilisateur a bien été supprimé");
-        header("Location: ../public/index.php?route=administration");
+        if ($this->checkAdmin()) {
+            $this->userDAO->deleteUser($userId);
+            $this->session->set("delete_user", "L'utilisateur a bien été supprimé");
+            header("Location: ../public/index.php?route=administration");
+        }
     }
 }
